# api/openapi.yaml
openapi: "3.0.3"
info:
  title: CRM Platform API
  version: 0.1.0
  description: Enterprise CRM platform with metadata-driven objects and granular security.

servers:
  - url: http://localhost:8080
    description: Local development

tags:
  - name: system
    description: System health and diagnostics
  - name: metadata-objects
    description: Object definition management
  - name: metadata-fields
    description: Field definition management

paths:
  /healthz:
    get:
      summary: Health check
      operationId: healthCheck
      tags:
        - system
      responses:
        "200":
          description: Service is healthy
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: ok

  /api/v1/admin/metadata/objects:
    post:
      summary: Create object definition
      operationId: createObject
      tags:
        - metadata-objects
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateObjectRequest"
      responses:
        "201":
          description: Object created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ObjectResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "409":
          $ref: "#/components/responses/Conflict"
        "500":
          $ref: "#/components/responses/InternalError"

    get:
      summary: List object definitions
      operationId: listObjects
      tags:
        - metadata-objects
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
            minimum: 1
        - name: per_page
          in: query
          schema:
            type: integer
            default: 20
            minimum: 1
            maximum: 100
        - name: object_type
          in: query
          schema:
            type: string
            enum: [standard, custom]
      responses:
        "200":
          description: List of objects
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ObjectListResponse"
        "500":
          $ref: "#/components/responses/InternalError"

  /api/v1/admin/metadata/objects/{objectId}:
    parameters:
      - $ref: "#/components/parameters/ObjectId"

    get:
      summary: Get object definition
      operationId: getObject
      tags:
        - metadata-objects
      responses:
        "200":
          description: Object definition
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ObjectResponse"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"

    put:
      summary: Update object definition
      operationId: updateObject
      tags:
        - metadata-objects
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateObjectRequest"
      responses:
        "200":
          description: Object updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ObjectResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"

    delete:
      summary: Delete object definition
      operationId: deleteObject
      tags:
        - metadata-objects
      responses:
        "204":
          description: Object deleted
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"

  /api/v1/admin/metadata/objects/{objectId}/fields:
    parameters:
      - $ref: "#/components/parameters/ObjectId"

    post:
      summary: Create field definition
      operationId: createField
      tags:
        - metadata-fields
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateFieldRequest"
      responses:
        "201":
          description: Field created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FieldResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "500":
          $ref: "#/components/responses/InternalError"

    get:
      summary: List field definitions for object
      operationId: listFields
      tags:
        - metadata-fields
      responses:
        "200":
          description: List of fields
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FieldListResponse"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"

  /api/v1/admin/metadata/objects/{objectId}/fields/{fieldId}:
    parameters:
      - $ref: "#/components/parameters/ObjectId"
      - $ref: "#/components/parameters/FieldId"

    get:
      summary: Get field definition
      operationId: getField
      tags:
        - metadata-fields
      responses:
        "200":
          description: Field definition
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FieldResponse"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"

    put:
      summary: Update field definition
      operationId: updateField
      tags:
        - metadata-fields
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateFieldRequest"
      responses:
        "200":
          description: Field updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FieldResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"

    delete:
      summary: Delete field definition
      operationId: deleteField
      tags:
        - metadata-fields
      responses:
        "204":
          description: Field deleted
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"

  # NOTE: Security endpoints (roles, permission-sets, profiles, users, groups,
  # sharing-rules) are registered manually via SecurityHandler.RegisterRoutes.
  # They are not included in OpenAPI codegen to avoid ServerInterface bloat.

components:
  parameters:
    ObjectId:
      name: objectId
      in: path
      required: true
      schema:
        type: string
        format: uuid

    FieldId:
      name: fieldId
      in: path
      required: true
      schema:
        type: string
        format: uuid

    Page:
      name: page
      in: query
      schema:
        type: integer
        default: 1
        minimum: 1

    PerPage:
      name: per_page
      in: query
      schema:
        type: integer
        default: 20
        minimum: 1
        maximum: 100

  schemas:
    CreateObjectRequest:
      type: object
      required:
        - api_name
        - label
        - plural_label
        - object_type
      properties:
        api_name:
          type: string
          minLength: 2
          maxLength: 100
          example: Invoice__c
        label:
          type: string
          minLength: 1
          example: Invoice
        plural_label:
          type: string
          minLength: 1
          example: Invoices
        description:
          type: string
          default: ""
        object_type:
          type: string
          enum: [standard, custom]
        is_visible_in_setup:
          type: boolean
          default: true
        is_custom_fields_allowed:
          type: boolean
          default: true
        is_deleteable_object:
          type: boolean
          default: true
        is_createable:
          type: boolean
          default: true
        is_updateable:
          type: boolean
          default: true
        is_deleteable:
          type: boolean
          default: true
        is_queryable:
          type: boolean
          default: true
        is_searchable:
          type: boolean
          default: false
        has_activities:
          type: boolean
          default: false
        has_notes:
          type: boolean
          default: false
        has_history_tracking:
          type: boolean
          default: false
        has_sharing_rules:
          type: boolean
          default: false
        visibility:
          type: string
          enum: [private, public_read, public_read_write, controlled_by_parent]
          default: private

    UpdateObjectRequest:
      type: object
      required:
        - label
        - plural_label
      properties:
        label:
          type: string
          minLength: 1
        plural_label:
          type: string
          minLength: 1
        description:
          type: string
        is_visible_in_setup:
          type: boolean
        is_custom_fields_allowed:
          type: boolean
        is_deleteable_object:
          type: boolean
        is_createable:
          type: boolean
        is_updateable:
          type: boolean
        is_deleteable:
          type: boolean
        is_queryable:
          type: boolean
        is_searchable:
          type: boolean
        has_activities:
          type: boolean
        has_notes:
          type: boolean
        has_history_tracking:
          type: boolean
        has_sharing_rules:
          type: boolean
        visibility:
          type: string
          enum: [private, public_read, public_read_write, controlled_by_parent]

    ObjectDefinition:
      type: object
      properties:
        id:
          type: string
          format: uuid
        api_name:
          type: string
        label:
          type: string
        plural_label:
          type: string
        description:
          type: string
        object_type:
          type: string
          enum: [standard, custom]
        is_platform_managed:
          type: boolean
        is_visible_in_setup:
          type: boolean
        is_custom_fields_allowed:
          type: boolean
        is_deleteable_object:
          type: boolean
        is_createable:
          type: boolean
        is_updateable:
          type: boolean
        is_deleteable:
          type: boolean
        is_queryable:
          type: boolean
        is_searchable:
          type: boolean
        has_activities:
          type: boolean
        has_notes:
          type: boolean
        has_history_tracking:
          type: boolean
        has_sharing_rules:
          type: boolean
        visibility:
          type: string
          enum: [private, public_read, public_read_write, controlled_by_parent]
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time

    ObjectResponse:
      type: object
      properties:
        data:
          $ref: "#/components/schemas/ObjectDefinition"

    ObjectListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/ObjectDefinition"
        pagination:
          $ref: "#/components/schemas/PaginationMeta"

    CreateFieldRequest:
      type: object
      required:
        - api_name
        - label
        - field_type
      properties:
        api_name:
          type: string
          minLength: 2
          maxLength: 100
        label:
          type: string
          minLength: 1
        description:
          type: string
          default: ""
        help_text:
          type: string
          default: ""
        field_type:
          type: string
          enum: [text, number, boolean, datetime, picklist, reference]
        field_subtype:
          type: string
          enum: [plain, area, rich, email, phone, url, integer, decimal, currency, percent, auto_number, date, datetime, time, single, multi, association, composition, polymorphic]
        referenced_object_id:
          type: string
          format: uuid
        is_required:
          type: boolean
          default: false
        is_unique:
          type: boolean
          default: false
        config:
          $ref: "#/components/schemas/FieldConfig"
        is_custom:
          type: boolean
          default: false
        sort_order:
          type: integer
          default: 0

    UpdateFieldRequest:
      type: object
      required:
        - label
      properties:
        label:
          type: string
          minLength: 1
        description:
          type: string
        help_text:
          type: string
        is_required:
          type: boolean
        is_unique:
          type: boolean
        config:
          $ref: "#/components/schemas/FieldConfig"
        sort_order:
          type: integer

    FieldConfig:
      type: object
      properties:
        max_length:
          type: integer
        precision:
          type: integer
        scale:
          type: integer
        format:
          type: string
        start_value:
          type: integer
        relationship_name:
          type: string
        on_delete:
          type: string
          enum: [set_null, cascade, restrict]
        is_reparentable:
          type: boolean
        default_value:
          type: string

    FieldDefinitionSchema:
      type: object
      properties:
        id:
          type: string
          format: uuid
        object_id:
          type: string
          format: uuid
        api_name:
          type: string
        label:
          type: string
        description:
          type: string
        help_text:
          type: string
        field_type:
          type: string
        field_subtype:
          type: string
        referenced_object_id:
          type: string
          format: uuid
        is_required:
          type: boolean
        is_unique:
          type: boolean
        config:
          $ref: "#/components/schemas/FieldConfig"
        is_system_field:
          type: boolean
        is_custom:
          type: boolean
        is_platform_managed:
          type: boolean
        sort_order:
          type: integer
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time

    FieldResponse:
      type: object
      properties:
        data:
          $ref: "#/components/schemas/FieldDefinitionSchema"

    FieldListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/FieldDefinitionSchema"

    PaginationMeta:
      type: object
      properties:
        page:
          type: integer
        per_page:
          type: integer
        total:
          type: integer
          format: int64
        total_pages:
          type: integer
          format: int64

    ErrorBody:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
        message:
          type: string

    ErrorResponse:
      type: object
      required:
        - error
      properties:
        error:
          $ref: "#/components/schemas/ErrorBody"

  responses:
    BadRequest:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Forbidden:
      description: Action not permitted
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Conflict:
      description: Duplicate or conflict
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"